+++ scripts/validate.sh
@@ -52,10 +52,13 @@ ansible-galaxy collection install "${validation_artifact}" --force -p "${validat
 
 export ANSIBLE_COLLECTIONS_PATH="${validation_install_dir}"
 for validation_module in \
+  minio_bucket \
   minio_user \
   minio_group \
+  minio_ldap_provider \
   minio_policy \
   minio_policy_binding \
+  minio_service \
   minio_service_account \
   minio_site_replication \
   minio_site_replication_info; do